Functions

Instanciation

TinyModia.@instantiateModel — Macro

modelInstance = @instantiateModel(model; FloatType = Float64, aliasReduction=true, unitless=false,
    log=false, logModel=false, logDetails=false, logStateSelection=false, logCode=false, logExecution=false, logTiming=false)

Instantiates a model, i.e. performs structural and symbolic transformations and generates a function for calculation of derivatives suitable for simulation.

model: model (declarations and equations)

FloatType: Variable type for floating point numbers, for example: Float64, Measurements{Float64}, StaticParticles{Float64,100}, Particles{Float64,2000}

aliasReduction: Perform alias elimination and remove singularities

unitless: Remove units (useful while debugging models and needed for MonteCarloMeasurements)

log: Log the different phases of translation

logModel: Log the variables and equations of the model

logDetails: Log internal data during the different phases of translation

logStateSelection: Log details during state selection

logCode: Log the generated code

logExecution: Log the execution of the generated code (useful for finding unit bugs)

logTiming: Log timing of different phases

return modelInstance prepared for simulation

TinyModia.instantiateModel — Function

See documentation of macro @instatiateModel

Simulation

TinyModia.simulate! — Function

simulate!(model [, algorithm];
          tolerance=1e-6, startTime=0.0, stopTime=1.0, interval=NaN, 
          adaptive=true, log=true, requiredFinalStates=nothing)

Simulate model::SimulationModel with algorithm (= alg of ODE Solvers of DifferentialEquations.jl). If the algorithm argument is missing, a default algorithm will be chosen from DifferentialEquations (for details see https://arxiv.org/pdf/1807.06430, Figure 3).

The simulation results stored in model can be plotted with ModiaPlot.plot and the result values can be retrieved with get_result.

Optional Arguments

tolerance: Relative tolerance.

startTime: Start time. If value is without unit, it is assumed to have unit [s].

stopTime: Stop time. If value is without unit, it is assumed to have unit [s].

interval: Interval to store result. If interval=NaN, it is internally selected as (stopTime-startTime)/500. If value is without unit, it is assumed to have unit [s].

adaptive: = true, if the algorithm should use step-size control (if available). = false, if the algorithm should use a fixed step-size of interval (if available).

log: = true, to log the simulation.

requiredFinalStates: is not nothing: Check whether the ODE state vector at the final time instant with @test is in agreement to vector requiredFinalStates with respect to some tolerance. If this is not the case, print the final state vector (so that it can be included with copy-and-paste in the simulate!(..) call).

Examples

import ModiaPlot
using  DifferentialEquations

# Runge-Kutta 5/4 with step-size control
simulate!(model, DifferentialEquations.Tsit5(), stopTime = 1.0)

    # Plot variables "v1", "v2" in diagram 1, "v3" in diagram 2, both diagrams in figure 3   
    ModiaPlot.plot(model, [("v1","v2"), "v3"], figure=3) 

    # Retrieve "time" and "v1" values:
    get_result(model, "time")
    get_result(model, "v1")

# Runge-Kutta 4 with fixed step size
simulate!(model, DifferentialEquations.RK4(), stopTime = 1.0, adaptive=false)

# Switching between Verners Runge-Kutta 6/5 algorithm if non-stiff region and 
# Rosenbrock 4 (= A-stable method) if stiff region with step-size control
simulate!(model, AutoVern6(Rodas4()), stopTime = 1.0)

# Automatically selected algorithm
simulate!(model, stopTime = 1.0)

# Sundials CVODE (BDF method with variable order 1-5) with step-size control
using Sundials
simulate!(model, CVODE_BDF(), stopTime = 1.0)

TinyModia.get_result — Function

signal = get_result(model, name; unit=true)

After a successful simulation of model::SimulationModel, return the result for the signal name::String as vector of points together with its unit. The time vector has path name "time". If unit=false, the signal is returned, without unit.

Example

using ModiaBase
using Unitful

include("$(ModiaBase.path)/demos/models/Model_Pendulum.jl")
using  .Model_Pendulum

pendulum = simulationModel(Pendulum)
simulate!(pendulum, stopTime=7.0)

time = get_result(pendulum, "time")  # vector with unit u"s"
phi  = get_result(pendulum, "phi")   # vector with unit u"rad"

import PyPlot
PyPlot.figure(4)   # Change to figure 4 (or create it, if it does not exist)
PyPlot.clf()       # Clear current figure
PyPlot.plot(ustrip(time), ustrip(phi), "b--", label="phi in " * string(unit(phi[1])))
PyPlot.xlabel("time in " * string(unit(time[1])))
PyPlot.legend()